Use PATCH instead of PUT for modify.
While it is arguable more proper in REST to have modify do a PUT,
in tmsh, and therefore iControl REST, modify is a PATCH.

For example, do a tmsh modify sys global-settings: only what you set is
modified. But do a PUT to /mgmt/tm/sys/global-settings and items you don't
set (mgmtDhcp, for example) will be set.